chore: refactor to a useAssetCaps context provider
