+++ src/components/transactions/Warnings/BorrowCapWarning.tsx
@@ -1,5 +1,5 @@
 import { Trans } from '@lingui/macro';
-import { AssetCapData } from 'src/hooks/getAssetCapUsage';
+import { AssetCapData } from 'src/hooks/useAssetCaps';
 import { Link } from '../../primitives/Link';
 import { Warning } from '../../primitives/Warning';
